Adecuaciones a segun version mas reciente del WSDL de AFIP

+++ src/fexv1/dif/afip/gov/ar/ClsFEXOutAuthorize.java
@@ -14,7 +14,7 @@ public class ClsFEXOutAuthorize  implements java.io.Serializable {
 
     private short cbte_tipo;
 
-    private short punto_vta;
+    private int punto_vta;
 
     private long cbte_nro;
 
@@ -37,7 +37,7 @@ public class ClsFEXOutAuthorize  implements java.io.Serializable {
            long id,
            long cuit,
            short cbte_tipo,
-           short punto_vta,
+           int punto_vta,
            long cbte_nro,
            java.lang.String cae,
            java.lang.String fch_venc_Cae,
@@ -124,7 +124,7 @@ public class ClsFEXOutAuthorize  implements java.io.Serializable {
      * 
      * @return punto_vta
      */
-    public short getPunto_vta() {
+    public int getPunto_vta() {
         return punto_vta;
     }
 
@@ -134,7 +134,7 @@ public class ClsFEXOutAuthorize  implements java.io.Serializable {
      * 
      * @param punto_vta
      */
-    public void setPunto_vta(short punto_vta) {
+    public void setPunto_vta(int punto_vta) {
         this.punto_vta = punto_vta;
     }
 
@@ -378,7 +378,7 @@ public class ClsFEXOutAuthorize  implements java.io.Serializable {
         elemField = new org.apache.axis.description.ElementDesc();
         elemField.setFieldName("punto_vta");
         elemField.setXmlName(new javax.xml.namespace.QName("http://ar.gov.afip.dif.fexv1/", "Punto_vta"));
-        elemField.setXmlType(new javax.xml.namespace.QName("http://www.w3.org/2001/XMLSchema", "short"));
+        elemField.setXmlType(new javax.xml.namespace.QName("http://www.w3.org/2001/XMLSchema", "int"));
         elemField.setNillable(false);
         typeDesc.addFieldDesc(elemField);
         elemField = new org.apache.axis.description.ElementDesc();
